Backend/requirements.txt
uvicorn
tensorflow
pillow
numpy
fastapi
python-multipart
